backend/utils/requirements.txt
Pillow>=10.0.0
opencv-python>=4.8.0
qrcode>=7.4.2
openpyxl>=3.1.0
numpy>=1.24.0
requests>=2.31.0
